requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
autopage==0.5.2
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.3.0
click==8.1.7
cliff==4.3.0
cmaes==0.10.0
cmd2==2.4.3
colorlog==6.7.0
contourpy==1.1.1
cycler==0.12.1
distlib==0.3.7
docker-pycreds==0.4.0
exceptiongroup==1.1.3
filelock==3.12.4
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.9.2
gitdb==4.0.10
gitpython==3.1.38
greenlet==3.0.0
hydra-colorlog==1.2.0
hydra-core==1.3.2
hydra-optuna-sweeper==1.2.0
identify==2.5.30
idna==3.4
imageio==2.31.5
imgaug==0.4.0
importlib-metadata==6.8.0
importlib-resources==6.1.0
iniconfig==2.0.0
jinja2==3.1.2
kiwisolver==1.4.5
lazy-loader==0.3
lightning-utilities==0.9.0
lightning==2.1.0
mako==1.2.4
markdown-it-py==3.0.0
markupsafe==2.1.3
matplotlib==3.8.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
nodeenv==1.8.0
numpy==1.22.3
omegaconf==2.3.0
opencv-python==4.8.1.78
optuna==2.10.1
packaging==23.2
pathtools==0.1.2
pbr==5.11.1
pillow==10.1.0
pip==23.3
platformdirs==3.11.0
pluggy==1.3.0
pre-commit==3.5.0
prettytable==3.9.0
protobuf==4.24.4
psutil==5.9.6
pygments==2.16.1
pyparsing==3.1.1
pyperclip==1.8.2
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytorch-lightning==2.1.0
pyyaml==6.0.1
requests==2.31.0
rich==13.6.0
rootutils==1.0.5
scikit-image==0.22.0
scipy==1.11.3
sentry-sdk==1.32.0
setproctitle==1.3.3
setuptools==68.0.0
shapely==2.0.2
six==1.16.0
smmap==5.0.1
sqlalchemy==2.0.22
stevedore==5.1.0
sympy==1.12
tifffile==2023.9.26
tomli==2.0.1
torch==2.1.0
torchmetrics==1.2.0
torchvision==0.16.0
tqdm==4.66.1
typing-extensions==4.8.0
urllib3==2.0.7
virtualenv==20.24.5
wandb==0.15.12
wcwidth==0.2.8
wheel==0.41.2
yarl==1.9.2
zipp==3.17.0